refactor(app): Update `dropTipInPlace` command text (#16610)
Closes EXEC-182

This commit updates tip drop command text when trash is dropped into a waste container, increasing specificity. Unfortunately, the command metadata itself does not give any location when performing a drop tip, so the app must derive the location from looking at the most recent addressableArea command. Fortunately, this should never be a compute-heavy scan unless the protocol does something like "move to trash" followed by a ton of "dispense over trash" before dropping tips, which seems quite unlikely in practice. Regardless, this instance serves as an example of places in which we should reduce compute done on the app in the future.