refactor(app):  Support "resume from recovery assuming false positive" action (#16601)
Closes EXEC-791

See #16556 and the linked ticket for details concerning the reasoning for implementing these changes.

When skipping a step in Error Recovery, the FE essentially calls two commands: handleIgnoringErrorKind, which deals with updating the recovery policy, and skipFailedCommand, which does the skipping. By using isAssumeFalsePositiveResumeKind, we can conditionally determine which policy ifMatch criteria to use, and we can determine which error recovery resume kind to dispatch to the server.