feat(api, robot-server, app): Handle the Plate Reader lid while held by the gripper when the Estop is pressed. (#16574)
We don't want customers manually handling the Plate Reader lid as this can cause damage to the device when you press the estop while the gripper is holding the plate reader lid it will eventually home the gantry and hold onto the lid. When a new protocol is started the gripper jaws will open, causing the plate reader lid to fall and potentially get damaged. This pull request makes it so we automatically place the plate reader lid in its dock (staging area) when the robot resumes operating after physically and logically disengaging the e-stop.

- Add a new `unsafe/placeLabware` command to PE to place labware the gripper is already holding on to.
- Add a new `placeLabwareState` to the `runs/{run_id}/currentState` endpoint to inform the client of the plate reader lid.
- Add a new `usePlacePlateReaderLid` hook on the client to conditionally call the `unsafe/placeLabware` command when the `placeLabwareState` returns the appropriate values.
- Fix an issue with the plate reader disconnecting when the estop was pressed
- Fix an issue where the plate reader lid was not loaded if the module was added during module setup.