fix(app): fix stale well location in Error Recovery (#16628)
Closes RQA-3438

The logic for determining the default well used during tip selection is calculated as soon as Error Recovery renders. After the initial well is determined, this state is set and then never updated for the remainder of Error Recovery. The current logic works well if there never has been a recentRelevantFailedLabwareCmd, however, if this is the 2nd+ instance of error recovery in the run, the recentRelevantFailedLabwareCmd is stale until the fetch for the newrecentRelevantFailedLabwareCmd completes.

Instead of gating the state setting behavior by whether or not there is a recentRelevantFailedLabwareCmd, we should gate it by whether the relevant data within recentRelevantFailedLabwareCmd has changed.