feat(api, robot-server): Add source key to the data_files database to support generated CSV files from the plate reader + other plate reader fixes. (#16603)
The Plate Reader generates absorbance readings that can be optionally saved as a CSV file on the robot with the
`read(export_filename='filename.csv')` PAPI command. This generated data needs to be stored on the robot so the client can download them later on for post-processing or re-upload them to be used as RTP CSV for a different protocol. Unlike RTP CSV that are uploaded by a user from the client, generated readings are produced by the robot. Because of this, we need to keep track of the source these data files originate from, so we can limit their count independently of RTP CSVs. A typical absorbance protocol makes about 40 readings per run, so we could hit our maximum number of data files in one run. This pull request adds a `source` key to the `data_files` table to keep track of the origin of the data files, changes to support the new functionality, and other fixes relating to the plate reader.

- Raise `CannotPerformModuleAction` analysis error if `initialize` is called before `close_lid`
- Add a `source` column to the `data_files` in the database to keep track of the origin of the data file.
- Use the `source` key to determine the number of uploaded files stored on the bot to prevent file creation.
- Use the `source` key to determine the number of generated files stored on the bot to prevent file creation.
- Increase the maximum number of generated CSV files to 400 to account for the average case of number of reads a plate reader could make per protocol run.