fix(app): Fix nested labware UI in Error Recovery (#16634)
Closes RQA-3440

This PR fixes copy and deck map issues when addressing labware nested in other labware.

5eab60c - Split the util that gets the labware location and renders appropriate copy into two separate utils.

c753bc5 - Fixes copy issues. We recently refactored slot location display utils, and I just forgot to use the correct, new prop here and update relevant i18n strings.

caf9340 - Fixes deck map issues. Essentially, if there are multiple labware in the slot with the failed labware, only render the topmost labware. This is consistent with deck map behavior in other places in the app. Yes, it's a pain and annoying to have to do this outside of the deck map, but after discussion with other FE devs, it sounds like the deck map redesign will fundamentally change the way we render all objects on the deck, so I don't think it's worth the effort to improve deck map internals currently.