fix(app): fix dropping tips in predefined waste chute location during Error Recovery (#16636)
Closes EXEC-794

8.2.0 adds predefined drop tip locations, and during drop tip wizard, the flow does not always have context concerning tip attachment status (ie, during maintenance runs). During a maintenance run, to automatically jog the pipette down somewhat before dropping the tip, we moveToAddressableArea with a predefined offset.

While this works well for maintenance run type drop tip flows, it does not always play nicely during fixit type drop flows (ie, during Error Recovery). When attempting to move to an addressable area with the predefined z-offset, the command may fail with an error pointing to the z-offset being invalid given the attached tip.

There are several potential ways to work around this including:

* Never specify an offset, always dropping into the waste chute (and trash bin) at maximum height. This works, but could be unideal in a real world scenario.
* Use the predefined z-offset during maintenance run type drop tip flows only (no offset during Error Recovery/fixit type flows). This works, however, the z-offset moved traversed during fixit type flows is less than during a maintenance run type flow.
* During a fixit type flow, get some information about currently attached tip length and use that.
To keep complexity down and reduce the bug surface, this PR implements option 2.

Additionally, it's probably not a good idea to drop tips or blowout if the moveToAddressableArea command fails, so let's change that!