feat(api,app,shared-data): Track air gaps (#16662)
We now are trying to track the amount of liquid in the wells of the
labware on the deck. To do this, we're tracking how much liquid we
aspirate from wells and dispense into wells.

Unfortunately, sometimes users want to aspirate air and then dispense it
into wells (well, they usually don't want to dispense it _into_ the
well, but they want to dispense it). Until now, we'd treat any volume of
air that got dispensed as a volume of liquid, and add it to the liquid
tracked in the well. This is wrong.

There's a couple things we need to do to fix this, and this PR does
them. It also does a couple other things to support that effort. It's
big, but well-separated into commits, and that's a decent way to do it.

- a6a0856823a0dfa49182a1e28f1df3fcc2fae02c adds a utility class that can
track the pipette contents and the relative ordering of air and liquid
slugs within the pipette. It does this so that you can say "Hey I'm
going to dispense x ul, how much of that is liquid?" and get a solid
answer.
- c19b205d29f6c8c9434bd1e3e0496b5aed68654a integrates that with existing
commands and state stores, in the process switching volume state
tracking to `StateUpdate`
- 7d81a3c0ae5be131826c62144df5a257b9fd86cc adds `airGapInPlace`, a new
command that works exactly like aspirate except that it doesn't take
volume out of a well, and it adds air to the pipette internal volume
tracking so that a subsequent dispense won't add liquid to the target
well
- 3f4d07197c17c79528c50998c11a693031fcd710 makes the python protocol API
use that for `air_gap` calls above a certain API version

There's also some supporting commits to add the new API version (since
this is going into edge) and a command text for the new command.

This is a pretty in depth thing and will need at least some testing, but
it is covered by pretty in-depth automated testing and I eagerly await
some failing analysis snapshots.

## reviewing
- can i get away with not changing the pipette state store tests to just
be about state updates? please?
- does the fluid stack thing seem overwrought, underwrought, wroughten
to the core?
- am I missing something?

## testing
- [x] inspect engine state that uses air gaps on api version 2.21. it
should still work
- [x] inspect engine state that uses air gaps on api version 2.22. it
should work in terms of physical actions
- [x] Inspect the volume trakcing and make sure the air isn't counted as
liquid

Closes EXEC-792 [@SyntaxColoring edit: also EXEC-754 incidentally, yay]

---------

Co-authored-by: TamarZanzouri <zanzouritamar@gmail.com>
Co-authored-by: Max Marrone <max@opentrons.com>