feat(api): add load-empty-liquid (#16676)
We have a loadLiquid command and associated python protocol API method,
which allow users to specify that a well contains a given amount of a
given (predefined) liquid. This was all that we needed for a while
because we didn't really do anything with that information except
display the starting deck state based on it.

Now that we track liquid, however, this isn't enough, because we also
need a way for users to specify that a well is known to be empty. To
that end, introduce a special EMPTY sentinel value for liquid ID and
restrict loadLiquid commands that use it to setting a 0 volume.

I think we probably should have a better API for setting multiple wells
but that's not this PR.

Closes EXEC-801

## reviews
- seem like a valid thing to do? Didn't make sense to have a whole
different command

## testing
- make sure empty liquids get loaded

## notes
- [x] will rebase this PR to edge once the current target is merged

---------

Co-authored-by: Ed Cormany <edward.cormany@opentrons.com>