fix(app): do not crash when you get a "C" locale (#16716)
Some linux systems are configured with one of the POSIX default locales,
like "C" or "POSIX", and for some reason this will be returned by
electron's getSystemPreferredLanguages() api directly. Unfortunately,
passing this to the browser js standard Intl.Locale() will make it
throw, and that would whitescreen the app. Instead, catch the error and
treat it as an unmatched system language.

Note that this crashes even if the feature flag isn't on.