refactor(api,robot-server,app): use labwareURI instead of labwareID for placeLabwareState and unsafe/placeLabware command. (#16719)
The `placeLabwareState` and `unsafe/placeLabware` commands rely on the
`labwareID` to place labware held by the gripper back down to a deck
slot. However, as correctly pointed out, this is wrong because the
labwareID changes across runs. This was working for the plate reader lid
because the lid is loaded with a fixed labwareID, however, that would not
be the case for other labware across different runs. This pull request
replaces the labwareID with labwareURI used by the `placeLabwareState`
and `unsafe/placeLabware` commands so the behavior is consistent across
runs.