feat(api): track volumes from multichannel configs (#16698)
This PR adds the capability to properly track volume changes made by
multichannel pipettes (and partial tip loadings of multichannel
pipettes) to the engine. It also adds a quick refactor of the visibility
of nozzle map types. This is well separated by commit.

## multichannels

This is in commit
https://github.com/Opentrons/opentrons/pull/16698/commits/d49e90b13fdcf51b95284b3349c5ac6f8446fb44

There are two ways in which we need to handle multichannel nozzle
configurations specially compared to single-channel configurations.

First, and what EXEC-795 is about, is that pipettes with multiple active
nozzles will aspirate out of or dispense into multiple wells in an
aspirate/dispense/in_place command. Which wells the pipette touches is a
matter of projecting the pipette nozzle map out over the layout of the
labware and predicting which wells are interacted with.

This is itself non-trivial because labware can have many formats. What
we can do is make the math work correctly when possible - when the
labware is laid out normally enough that we can do projections of this
type - and fall back to pretending to be a single channel if we fail.
Since we're computing the logical equivalent of actual physical state,
and if the labware is irregular it's unlikely that a multiple nozzle
layout will physically work with the labware, I think this is safe.

Specifically the thing we need to do is generalize the logic used in the
tip store to project which tips are picked up by a multichannel to
labware of different formats. Our multichannel pipette nozzles are laid
out to match SBS 96-well plates, and so that's our "default" labware. On
labware that follows SBS patterns but is more dense - a 384 plate, for
instance - then we have to subsample, picking a single well in each
group of (well_count / 96) that occupies the same space as a 96-well
well to interact with. On labware that follows SBS patterns but is less
dense - a 12-column reservoir, for instance - then we have to
supersample, letting a labware well be touched by multiple nozzles.

The second thing we have to deal with is that if the labware is a
reservoir or reservoir-like - it has fewer wells than we have nozzles -
then the common case is that multiple nozzles are in a well, and in that
case if we're keeping track of the volume taken out of or added into a
well we have to multiply the operation volume by the number of nozzles
per well, which we can get by just dividing sizes without taking into
account pattern overlap.

## nozzle maps

This is in commit
https://github.com/Opentrons/opentrons/pull/16698/commits/d64b929fd8a0403031c36250f8486081018d9f5d

This came up as I was poking around with needing the nozzle map to be
visible in new places; I found it pretty awful that it was just
implicitly exposed including internals, so make a new interface protocol
that is explicitly exposed in `opentrons.types` and hold all of the
internals, well, internal, at least to the engine.

Closes EXEC-795

## to come out of draft
- [x] tests. lots of tests
- [x] general approval that this is the way forward

## testing
`opentrons_cli analyze` is probably enough for these because it's all
logical state manipulation.
- [x] this works for a single pipette the way it has always works, ditto
1-channel configurations on 8 or 96 channel pipettes
- [x] 8-channel pipettes properly handle 96-standard full-column
operations
- [x] 8-channel pipettes properly handle 96-standard offset operations
- [x] 8-channel pipettes handle 384 plates, including A1 and B1
operations
- [x] 8-channel pipettes handle 12-column reservoirs
- [x] 96-channel pipettes handle 96-standard full operations
- [x] 96-channel pipettes handle 384 plates, including A1, B1, A2, and
B2 operations
- [x] 96-channel pipettes handle 12-column and 1 column reservoirs

---------

Co-authored-by: Ryan Howard <ryan.howard@opentrons.com>