fix(app): fix ODD IntersectionObserver reference cycle (#16743)
Works toward EXEC-807

We have this code in the ODD that runs every render cycle on a lot of the "idle" views that manages the scrollbar. We instantiate a new observer on every render, but the old observer is never cleaned up. This commit ensures that we only ever instantiate one observer, and we properly remove it on component derender.