feat(api): implement a store to save liquid classes in the Protocol Engine (#16732)
# Overview

First step of AUTH-851.

Define a `LiquidClassStore` in the Protocol Engine that stores liquid
classes. Following the usual pattern, the store contains a
`LiquidClassState`, and you update the state by sending a
`LiquidClassLoadedUpdate`, and there's a `LiquidClassView` that provides
a read-only view of the store, etc.

The entity that we're storing in the store is a `LiquidClassRecord`,
which consists of:
- the liquid class name,
- the pipette ID,
- the tiprack ID,
- the transfer properties, which consists of `AspirateProperties`,
`SingleDispenseProperties`, `MultiDispenseProperties`, plus any future
properties we might introduce.

To avoid having to duplicate code, the `LiquidClassRecord` just inherits
from `ByTipTypeSetting`, which contains the tiprack ID and the transfer
properties.

Inside the Protocol Engine, liquid classes are *immutable*. When the
client sends a liquid class to the Protocol Engine, we have to store it
under a new ID if the liquid class is different from any
`LiquidClassRecord` we've already stored. So the most interesting part
of this PR is implementing lookups by `LiquidClassRecord` values:

The most natural way to implement lookups is with a `dict`, which
requires hashing. But Pydantic will not generate a usable hash function
if any of the subfields are not frozen. And it would be too invasive to
make every single subfield (like `Coordinate`) frozen. I tried using
packages like `frozendict` that would implement the hashing for us, but
it was also too much work to bring in a new package. So I ended up
having to implement the hash function on `LiquidClassRecord` manually.

## Test Plan and Hands on Testing

Added unit tests for the liquid class store and view.

In particular, I verified that the hash implementation can successfully
query for `LiquidClassRecord` values.

## Review requests

This is my first time touching the state in the Protocol Engine, so let
me know if I'm missing something.

This is just the first part; the `loadLiquidClass` command
implementation will be in a separate PR.

## Risk assessment

Low risk. Liquid classes are currently don't interact with any other
aspect of the Protocol Engine, and this change should not have any
externally visible effects.