fix(protocol-designer): fix padding and footer position in onboard wizard (#16832)
This PR fixes two bugs in the onboarding wizard. Here, I set a minimum
grid gap between the left side content and footer buttons so that the
buttons don't reposition on steps of different heights. I also fix the
padding for the left side content to be a constant 5rem.

Closes RQA-3512