feat(api): implement loadLiquidClass command in PE (#16814)
# Overview

This is the second half of AUTH-851.

This implements the `loadLiquidClass()` command using the liquid class
store from the previous PR.

Liquid classes are read-only in the Protocol Engine, so each
`liquidClassId` refers to one specific liquid class definition. Each
mutation of a liquid class needs to be stored under a different
`liquidClassId`. The caller can specify the `liquidClassId` if they want
to, or else we will generate one for them.

So there are 4 cases we need to handle:

1. Caller did not specify a `liquidClassId` and the liquid class is new:
Generate a new `liquidClassId`.
2. Caller did not specify a `liquidClassId` but the liquid class has
already been stored: Reuse the existing `liquidClassId`.
3. Caller specified a `liquidClassId` that we haven't seen before: Store
the liquid class under the new `liquidClassId`.
4. Caller specified a `liquidClassId` that's already been loaded: Check
that the incoming liquid class matches the one we previously stored,
and:
    a. If the incoming liquid class exactly matches the existing one, do
nothing.
    b. If they don't match, raise an error.

## Test Plan and Hands on Testing

I added 5 test cases to cover each of the scenarios above.

## Review requests

Could someone teach me what `StateView.get_summary()` is supposed to do?

## Risk assessment

Low risk, liquid classes are not released yet and these changes should
be dev-only. This change makes the Protocol Engine state marginally
larger.